Move non pages out of pages folder
Rename User component to EditProfile in own folder
Move user page subfolders to components
Move UserInfoContext to context folder
Fix imports in all files
