+++ identity-admin/webapp/components/AccountActions/AccountActionButton.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button } from '../../../components/Button';
+import { Button } from '../Button';
 
 type AccountActionButtonProps = {
   label: string;
