+++ identity-admin/webapp/components/EditProfile/index.ts
@@ -1,0 +1,1 @@
+export { EditProfile } from './EditProfile';
