+++ identity-admin/webapp/pages/user/[userId].tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { withPageAuthRequired } from '@auth0/nextjs-auth0';
-import User from './User';
-import { UserInfoProvider } from './UserInfoContext';
+import { EditProfile } from '../../components/EditProfile';
+import { UserInfoProvider } from '../../context/UserInfoContext';
 
 function UserPage(): JSX.Element {
   return (
     <UserInfoProvider>
-      <User />
+      <EditProfile />
     </UserInfoProvider>
   );
 }
